This commit adds the review card to the project
