+++ cards/db.json
@@ -2,30 +2,98 @@
     "reviews":[
         {
             "id" : 1,
-            "name": "Perry",
+            "name": "berry",
             "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
             "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
             "content" : "I would recommend this spa anytime"
         },
         {
-            "id" : 1,
-            "name": "Perry",
+            "id" : 2,
+            "name": "Petty",
             "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
             "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
             "content" : "I would recommend this spa anytime"
         },
         {
-            "id" : 1,
+            "id" : 3,
             "name": "Perry",
             "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
             "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
             "content" : "I would recommend this spa anytime"
         },
         {
-            "id" : 1,
+            "id" : 4,
+            "name": "betty",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 5,
             "name": "Perry",
             "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
             "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 6,
+            "name": "Petty",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 7,
+            "name": "berry",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 8,
+            "name": "Petty",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 9,
+            "name": "Perry",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 10,
+            "name": "beky",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 11,
+            "name": "Perry",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
+            "content" : "I would recommend this spa anytime"
+        },
+        {
+            "id" : 12,
+            "name": "Petty",
+            "imageurl" : "https://i.bandori.party/u/c/art/a/tthumb/1301Maya-Yamato-Happy-VNzmri.png",
+            "title" : "great place",
+            "rating": "⭐⭐⭐⭐⭐",
             "content" : "I would recommend this spa anytime"
         }
     ]
